Post 목록 관리하는 리사이클러뷰 어댑터 코드 너무길어서 따로 파일로 빼면서 패키지도 생성. 픽셀, dp 변환 함수도 클래스빼면서 static 함수로
