+++ app/src/main/res/layout/list_item_post.xml
@@ -21,8 +21,7 @@
             <RelativeLayout
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:layout_marginLeft="4dp"
-                android:layout_marginRight="4dp">
+                android:layout_marginLeft="8dp">
 
                 <TextView
                     android:layout_width="wrap_content"
